requirements.txt
fastmcp>=0.1.0
crawl4ai>=0.1.0
pydantic>=2.0.0
requests>=2.28.0
curl-cffi>=0.5.5